Updated EnemyAI
Add back old models and json for testing purposes. Enemies now spawn and take formation in front of the boss and begin firing at the player. Firing currently is continuous and in a straight line until more logic has been decided for each type of enemy and how they will combat the player.