Updated enemy attacks
Added orb attack for use with the satellite boss. Updated the laser attack in the ini file to where only one entry existed. Added in functions to help spawn in the shield portion of final boss depending on how it is used. Boss behavior three has been updated to begin laying ground work of final boss.